requirements.txt
apache-airflow==2.5.0
apache-airflow-providers-amazon==3.3.0
boto3==1.24.0
numpy==1.23.0
pyyaml==6.0
logging==0.5.1.2
apache-airflow-providers-postgres==5.6.0
psycopg2-binary==2.9.7
requests==2.31.0
flask==2.3.3
celery==5.3.1
redis==4.6.0